Create NewTask_handler
